Tests/LibWeb/Text/expected/css/style-invalidation/active-state-featureless-has-targeted-invalidation.txt

active color: rgb(1, 2, 3)
FAIL: :active state change recomputed 509 elements
